Generate code offline (not in build.rs)
The codegen step depends on cargo-expand being installed. docs.rs did
not like it being invoked from a build.rs file (the build would fail)
and additionally, sqltk could not be published to crates.io without a
--no-verify flag.

The approach taken now is to generate code offline (which will be
checked in to the repo).

To run the codegen using this command in the root of the workspace:

`cargo run -p sqltk-codegen`

It writes generated code to sqltk/src/generated.
