+++ packages/sqltk-codegen/src/generics.rs
