+++ packages/sqltk-codegen/src/reachability.rs
